requirements.txt
flask>=3.0.0
pymysql>=1.1.0
cryptography>=42.0.0
opencv-python-headless>=4.9.0
numpy>=1.26.0
Pillow>=10.0.0
Werkzeug>=3.0.0
tensorflow>=2.16.0
